Updated sensors to use floats. Added new Sensor Entry and removed sensor packet.

+++ src/Sannel.House/Sensor/IHumiditySensor.cs
@@ -20,6 +20,6 @@ namespace Sannel.House.Sensor
 	[Exportable(Includes = @"#include ""ISensor.h""")]
 	public interface IHumiditySensor : ISensor
 	{
-		double GetRelativeHumidity();
+		float GetRelativeHumidity();
 	}
 }
